+++ picorouter/__init__.py
@@ -1,2 +1,2 @@
 # PicoRouter
-__version__ = "0.0.3"
+__version__ = "0.0.3c"